implement message handler for the MyActor
a simple handler is implemented that when sent a Message with a Shape,
reponds with the same dummy answer with a draw outcome.

but it reponds with a Future that's actually resolved after a delay of 3
seconds.

this is to simulate how it would work in the end.

the webapp resource handler has been updated to call (send a message to)
the Actor, and map the result in a HttpResponse.